backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Marsel <lllmarselll@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
email-validator = "^2.2.0"
alembic = "^1.14.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "4.0.1"
types-passlib = "^1.7.7.20240819"
types-python-jose = "^3.3.4.20240106"
python-multipart = "^0.0.17"
pytz = "^2024.2"
types-pytz = "^2024.2.0.20241003"
celery = "^5.4.0"
redis = "^5.2.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
